+++ info/index.php
@@ -2,7 +2,12 @@
 namespace ConnectFour;
 
 if (!defined('ROOT')) define('ROOT', dirname(__DIR__).'/');
-require_once(ROOT.'lib.php');
+require_once ROOT.'lib.php';
 
-$info = ['width' => BOARD_WIDTH, 'height' => BOARD_HEIGHT, 'strategies' => STRATEGIES];
+$info = [
+    'width' => BOARD_WIDTH,
+    'height' => BOARD_HEIGHT,
+    'strategies' => STRATEGIES,
+];
+
 Lib\response($info);