+++ model/board.php
@@ -2,7 +2,7 @@
 namespace ConnectFour\Model;
 
 if (!defined('ROOT')) define('ROOT', dirname(__DIR__) . '/');
-require_once(ROOT.'lib.php');
+require_once ROOT.'lib.php';
 
 class Board implements \JsonSerializable
 {
@@ -36,25 +36,17 @@ class Board implements \JsonSerializable
 
     public function doMove($move, $player)
     {
-        assert($this->columnHeights[$move] < BOARD_HEIGHT, "Column $move is full.");
+        assert($this->columnHeights[$move] < BOARD_HEIGHT,
+            "Column $move is full.");
 
         $this->columns[$move] |= $player << $this->columnHeights[$move];
         $this->columnHeights[$move]++;
     }
 
-    public function undoMove($move)
-    {
-        assert($this->columnHeights[$move] > 0, "Column $move is empty.");
-
-        $this->columnHeights[$move]--;
-        $player = $this->columns[$move] >> $this->columnHeights[$move] & 1;
-        $this->columns[$move] &= (1 << $this->columnHeights[$move] + 1) - 1;
-        return $player;
-    }
-
     public function getDiskOwner($column, $row)
     {
-        assert($column >= 0 && $column < BOARD_WIDTH && $row >= 0 && $row < BOARD_HEIGHT, "Invalid board position, ($column, $row).");
+        assert($column >= 0 && $column < BOARD_WIDTH && $row >= 0 && $row < BOARD_HEIGHT,
+            "Invalid board position, ($column, $row).");
 
         if ($this->columnHeights[$column] <= $row) {
             // No disk at position
@@ -93,28 +85,29 @@ class Board implements \JsonSerializable
         $center = [$column, $row];
 
         //Check vertical
-        $group = array_merge($center, $this->getRestOfGroup($player, $column, $row, 0, -1));
+        $group = array_merge($center, $this->getRestOfGroup($player, $column,
+            $row, 0, -1));
         if (count($group) >= 8) {
             return $group;
         }
 
         //Check Horizontal
-        $group = array_merge($center, $this->getRestOfGroup($player, $column, $row, -1, 0),
-                             $this->getRestOfGroup($player, $column, $row, 1, 0));
+        $group = array_merge($center, $this->getRestOfGroup($player, $column,
+            $row, -1, 0), $this->getRestOfGroup($player, $column, $row, 1, 0));
         if (count($group) >= 8) {
             return $group;
         }
 
         //Check Diagonal 1
-        $group = array_merge($center, $this->getRestOfGroup($player, $column, $row, 1, 1),
-                             $this->getRestOfGroup($player, $column, $row, -1, -1));
+        $group = array_merge($center, $this->getRestOfGroup($player, $column,
+            $row, 1, 1), $this->getRestOfGroup($player, $column, $row, -1, -1));
         if (count($group) >= 8) {
             return $group;
         }
 
         //Check Diagonal 2
-        $group = array_merge($center, $this->getRestOfGroup($player, $column, $row, 1, -1),
-                             $this->getRestOfGroup($player, $column, $row, -1, 1));
+        $group = array_merge($center, $this->getRestOfGroup($player, $column,
+            $row, 1, -1), $this->getRestOfGroup($player, $column, $row, -1, 1));
         if (count($group) >= 8) {
             return $group;
         }
@@ -124,6 +117,9 @@ class Board implements \JsonSerializable
 
     public function jsonSerialize()
     {
-        return ['columns'=>$this->columns, 'columnHeights'=>$this->columnHeights];
+        return [
+            'columns'=>$this->columns,
+            'columnHeights'=>$this->columnHeights,
+        ];
     }
 }